Convert to float

+++ src/core/macros.h
@@ -1,4 +1,4 @@
 #pragma once
 
 #define maximum(a, b) ((a) > (b) ? (a) : (b))
-#define round(x) ((int)((x) + 0.5))
+#define round(x) ((float)((int)((x) + 0.5f)))
